+++ cx-portal/src/components/pages/Admin/components/RegistrationRequests/CompanyDetailOverlay/index.tsx
@@ -107,12 +107,13 @@ const CompanyDetailOverlay = ({
 
   const downloadDocumnet = async (documentId: string, documentType: string) => {
     try {
-      const response = await getDocumentById(documentId).unwrap()
+      if (company) {
+        const response = await getDocumentById({ appId: company.applicationId, documentId }).unwrap()
+        const fileType = response.headers.get('content-type')
+        const file = response.data
 
-      const fileType = response.headers.get('content-type')
-      const file = response.data
-
-      return download(file, fileType, documentType)
+        return download(file, fileType, documentType)
+      }
     } catch (error) {
       console.error(error, 'ERROR WHILE FETCHING DOCUMENT')
     }
@@ -280,18 +281,16 @@ const CompanyDetailOverlay = ({
                     key="Street"
                     {...{
                       variableName: 'Street',
-                      value: `${selectedCompany?.streetName || ''} ${
-                        selectedCompany?.streetNumber || ''
-                      }`,
+                      value: `${selectedCompany?.streetName || ''} ${selectedCompany?.streetNumber || ''
+                        }`,
                     }}
                   />
                   <DetailGridRow
                     key="PLZ / City"
                     {...{
                       variableName: 'PLZ / City',
-                      value: `${selectedCompany?.zipCode || ''} ${
-                        selectedCompany?.city || ''
-                      }`,
+                      value: `${selectedCompany?.zipCode || ''} ${selectedCompany?.city || ''
+                        }`,
                     }}
                   />
                   <DetailGridRow
@@ -381,7 +380,7 @@ const CompanyDetailOverlay = ({
                       </Typography>
                     </Grid>
                     {selectedCompany?.companyRoles &&
-                    selectedCompany?.companyRoles.length > 0 ? (
+                      selectedCompany?.companyRoles.length > 0 ? (
                       <>
                         {selectedCompany?.companyRoles.map(
                           (role: { companyRole: string }) => (
@@ -401,8 +400,8 @@ const CompanyDetailOverlay = ({
                                   margin: '0px 10px',
                                   cursor: 'auto',
                                 }}
-                                onClick={function noRefCheck() {}}
-                                onFocusVisible={function noRefCheck() {}}
+                                onClick={function noRefCheck() { }}
+                                onFocusVisible={function noRefCheck() { }}
                                 size="small"
                                 variant="contained"
                               >